# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/api/auth/users": {
            "post": {
                "tags": [
                    "Auth user"
                ],
                "summary": "Create new user",
                "parameters": [
                    {
                        "description": "User",
                        "name": "string",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.UserRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
...
        "/api/session": {
            "post": {
                "tags": [
                    "Auth user"
                ],
                "summary": "Login user",
                "parameters": [
                    {
                        "description": "User",
                        "name": "string",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.UserRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
        "/api/session/logout": {
            "post": {
                "tags": [
                    "Auth user"
                ],
                "summary": "Logout user",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        }
    },
    "definitions": {
...
        "model.UserRequest": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string",
                    "example": "johndoe@mail.com"
                },
                "password": {
                    "type": "string",
                    "example": "qwerty"
                }
            }
        }
    }
}